+++ lib/services/user_settings.dart
@@ -5,11 +5,14 @@ abstract class UserSettings {
   Future<void> setDarkMode(bool value);
   Future<int?> getBuildNumber();
   Future<void> setBuildNumber(int buildNumber);
+  Future<int?> getNewVerseFrequency();
+  Future<void> setNewVerseFrequency(int frequency);
 }
 
 class SharedPreferencesLocalStorage extends UserSettings {
   static const String _darkModeKey = 'darkMode';
   static const String _buildNumberKey = 'buildNumber';
+  static const String _newVerseFrequencyKey = 'newVerseFrequency';
 
   @override
   Future<bool> getDarkMode() async {
@@ -33,5 +36,17 @@ class SharedPreferencesLocalStorage extends UserSettings {
   Future<void> setBuildNumber(int buildNumber) async {
     final prefs = await SharedPreferences.getInstance();
     await prefs.setInt(_buildNumberKey, buildNumber);
+  }
+
+  @override
+  Future<int?> getNewVerseFrequency() async {
+    final prefs = await SharedPreferences.getInstance();
+    return prefs.getInt(_newVerseFrequencyKey);
+  }
+
+  @override
+  Future<void> setNewVerseFrequency(int frequency) async {
+    final prefs = await SharedPreferences.getInstance();
+    await prefs.setInt(_newVerseFrequencyKey, frequency);
   }
 }